chore: rename package to easi encoding
