+++ pkg/easiencoding/base_64.go
@@ -1,4 +1,4 @@
-package easiencryption
+package easiencoding
 
 import (
 	"bytes"
