+++ pkg/graph/resolvers/system_intake_document.go
@@ -8,7 +8,7 @@ import (
 	"github.com/google/uuid"
 
 	"github.com/cmsgov/easi-app/pkg/appcontext"
-	"github.com/cmsgov/easi-app/pkg/easiencryption"
+	"github.com/cmsgov/easi-app/pkg/easiencoding"
 	"github.com/cmsgov/easi-app/pkg/graph/model"
 	"github.com/cmsgov/easi-app/pkg/models"
 	"github.com/cmsgov/easi-app/pkg/storage"
@@ -59,7 +59,7 @@ func CreateSystemIntakeDocument(ctx context.Context, store *storage.Store, s3Cli
 		s3Key += fallbackExtension
 	}
 
-	decodedReadSeeker, err := easiencryption.DecodeBase64File(&input.FileData.File)
+	decodedReadSeeker, err := easiencoding.DecodeBase64File(&input.FileData.File)
 	if err != nil {
 		return nil, fmt.Errorf("...%w...FileName: %s", err, input.FileData.Filename) //Wrap error and provide filename
 	}
