+++ pkg/graph/resolvers/trb_request_document_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/99designs/gqlgen/graphql"
 	"github.com/google/uuid"
 
-	"github.com/cmsgov/easi-app/pkg/easiencryption"
+	"github.com/cmsgov/easi-app/pkg/easiencoding"
 	"github.com/cmsgov/easi-app/pkg/graph/model"
 	"github.com/cmsgov/easi-app/pkg/models"
 )
@@ -36,7 +36,7 @@ func (suite *ResolverSuite) TestTRBRequestDocumentResolvers() {
 
 func createTRBRequestDocumentSubtest(suite *ResolverSuite, trbRequestID uuid.UUID, documentToCreate *models.TRBRequestDocument) *models.TRBRequestDocument {
 	testContents := "Test file content"
-	encodedFileContent := easiencryption.EncodeBase64String(testContents)
+	encodedFileContent := easiencoding.EncodeBase64String(testContents)
 	fileToUpload := bytes.NewReader([]byte(encodedFileContent))
 	gqlInput := model.CreateTRBRequestDocumentInput{
 		RequestID:            documentToCreate.TRBRequestID,
